Patch contributed by Paul Moore <paulmoore100@hotmail.com>
The purpose of this patch set is to implement GSS authentication -
particularly useful when talking to Microsoft's winrm stack. Here it is in
action.


C:\Documents and Settings\Administrator.ENT2K30>winrm identify
-r:http://paul-es5:8889/wsman -auth:Kerberos
IdentifyResponse
    ProtocolVersion = http://schemas.dmtf.org/wbem/wsman/1/wsman.xsd
    ProductVendor = Openwsman Project
    ProductVersion = 2.0.0

I get a secure (mutually authenticated and encrypted) session without certs
and without user / password.

In the course of adding GSS I came across other issues in the base http
handling. The fixes for these are not ifdef'd; the changes that imlement GSS
are ifdef'd. Actually the ifdef is probably not needed since openwsman seems
to require kerb libs and headers to build anyway (for curl lib I think). Up
to you whether you want to keep them ifdef's or not. I did not change
configure (I don't have your configure.ac) to make it an option. My vote
would be to un-ifdef it

Non GSS specifc issues

A) ws_base64_encode
Doesn't work at all. 
GSS sends base64 encoded data back to the client so this is why I needed it

B) keep-alive
Did not work. Needed by GSS because it establishes a context for a
connection, the requests are not atomic, the actual real request is never
available until the second HTTP request on a given TCP pipe.
Main fix was resetting the 'headers have been parsed' flag on the connection


C) large output buffers
GSS generates big messages. Wsmand-listener has code to send a message in
multiple chunks. This did not work correctly
See fixes near the label 'CONTINUE:' in wsmand-listener

D) changed timeout value when shttpd_poll called