fixed possible segfault (bug 272)
In the file sfcc-interface.c; character pointer element is declared as:
"char* element;" on line 1583 and is not initialized till line 1611.


However, if the code falls through the if conditional on line 1606; the
cleanup code is called on 1609 and the cleanup code will try to free up
element (line 1696), while it might still be pointing to some random
memory location.


Following is the patch to fix this problem:


Index: src/plugins/cim/sfcc-interface.c
===================================================================
--- src/plugins/cim/sfcc-interface.c    (revision 3016)
+++ src/plugins/cim/sfcc-interface.c    (working copy)
@@ -1580,7 +1580,7 @@
        WsmanStatus statusP;
        CMPIConstClass *class;
        int fragment_flag;
-       char *element;
+       char *element = NULL;
        int index;
        WsXmlNodeH resource;
        WsXmlNodeH child = NULL;

suresh sundriyal