From Chris_Poblete@Dell.com
Certain WSMAN request causes a segmentation fault crash on an Apache
thread processing requests to OpenWSMAN via mod_wsman. The crash is
occurring consistently from an XML request that contains a value for
"Option" as it appears below:
  <s:Header>
  ...
  <w:OptionSet>
    <w:Option Name="ShowExtensions">true</w:Option>
    </w:OptionSet>
    ...
  </s:Header>
  
Core dump stack trace shows crash on call to malloc from Apache
suggesting heap corruption.

Examination of the code found that a call to get the value of "Option"
from an XML with a value returns a (const) reference that is explicitly
freed by the caller.