Send http/1.1 "connection: close" as part of response
In the file src/server/wsmand-listener.c in function server_callback()
after doing all processing on the WSMAN request we send a response
back to the client. The response can be anything ( a valid WSMAN
response or a Fault), but we send some thing.

Here in OpenWSMAN we are not having any session based connection.
Every request is a new connection. So when we +send a response back we
close the connection and we need to inform the client that connection
needs to be closed.

(Patch by Arun Venkatachalam <A_Venkatachalam@Dell.com>