Fix escaping of XML data send from sfcb
Found and fixed by A_Venkatachalam@Dell.com:

There is a discrepancy in principle the way OpenWSMAN is treating the
data coming from SFCB i.e. payload while using the CIM-XML and Local
interface of SFCB. The issue is the way in which payload that comes
when we have XML as payload.

a.) Under the CIM XML mechanism SFCB encodes the XML payload or any
    payload that is coming inside CIM XML headers
b.) Under Local interface as SFCB expects that its not putting data on
    the wire I does not encode the XML payload

Note: We are assuming that the SFCB providers are returning data as
XML and not as plain strings, but above is applicable to strings also
which can have special characters and SFCB will do an
encoding/escaping of them.

So the fundamental  question is -- Should OpenWSMAN try to decode the
data under both situations a. and b. ? Currently OpenWSMAN is doing
the same in both cases -- It decodes the data.
An example is as follows -

1. Lets us assume SFCB provider is sending XML data back via SFCB
   server and XML data is already escaped (encoded) it. i.e. '&' will be
   represented as &amp;
2. Now when SFCB picks it up and gives it back to OpenWSMAN
        a. We know it encodes so &amp; will become &amp;ampl; (double
           escaping) -- In case of CIMXML
        b. But in case if local interface it will not encode -- so
           data will go to OpenWSMAN as &amp; itself.
3. So if we observe In case a. The way data reaches OpenWSMAN is
   different from the way Data reaches in case of b.

Currently in OpenWSMAN 2.1 while working on the data received
OpenWSMAN decodes the data and works on it in both the cases, this it
does by making use of xmlNodeSetContent() API of libwsman which
internally does a decoding. This is very much valid in case a. but in
case of b. it should not do this because if it does this, &amp; which
was the original data that was not encoded in local interface will
become '&' which should not happen as XML data that was received in
SFCB was &amp; and not '&' and that’s how OpenWSMAN should give back
to the external world.

Ultimately, when the data is thrown out by OpenWSMAN back to the wire
for its clients to utilize the data will be different in both the
cases as OpenWSMAN Solution --- In case of local interface we must
make use of Libxml's xmlNodeAddContent() API as it expects raw data as
input and does not do an decoding and sends the data as it is.