Code cleanup (possible mem leaks and NULL references) in cim_data_stubs.c
Found and fixed by ssundriy@vmware.com


1. Possible Segfault:

In method CimResource_Init(), there should be a check for allocation
failure right after the following line:

CimClientInfo *cimclient=
(CimClientInfo*)u_zalloc(sizeof(CimClientInfo));

Considering we dereference cimclient->cc a couple of lines later, if
there is an allocation failure, this will lead to a segfault. The
included patch has a fix for this.

2. Memory Leak:

Now considering the case where the allocation has succeeded, if
initialization of cimclient->cc fails, we return with a NULL without
cleaning up the space allocated for cimclient.  This would definitely be
a memory leak. The attached patch has a fix for this as well.

3. Another Segfault:

If you look at CimResource_Delete_EP, we initialize cimclient only if
msg is not NULL and there is no else clause. This means if we get a case
where msg is NULL, cimclient is not initialized and passed as NULL to
verify_class_namespace, which does some funky stuff with it which is
guaranteed to segfault if we hit this corner case.

CimResource_Delete_EP is not the only culprit here, most of the endpoint
callback methods in this file suffer from this affliction.

The patch adds code to verify_class_namespace that checks whether the
passed in pointer is null or not and returns '0' if it is, which seems
to mitigate some of the problems associated with the enpoint callback
methods(more on that later).

4. Even more segfaults:

In verify_class_namespace, even if the *client is initialized, looking
at CimResource_Init, it seems that client->resource_uri,
client->requested_class and client->method can still be NULL. We pass
these potentially NULL pointers to strstr() and strcmp() which will
again cause a segfault. Not sure under what circumstances this will
happen but the code certainly has no provision to accommodate these
circumstances.