While testing indication persistence, I came across a weird bug. Every time I restarted  Openwsman the expiretime on the saved subscriptions would increase by an hour. This took a while to figure out.
In ws_set_expiretime, we use mktime to convert tmx.tm structure to convert to epoch
time. mktime() corrects for dst based on the value of tmx.tm.tm_isdst.
However, in ws_serialize_datetime() method, where we actually populate tmx.tm
structure, we do not populate the value of tmx.tm.tm_isdst. This causes a time skew on
systems that have the value set.

The patch fixes the code so that the tmx.tm.tm_isdst structure is populated while
serializing xmldatetime.

Found and fixed by Suresh Sundriyal