Suresh Sundriyal writes:
I'm sending along a patch for the problem above. It seems that when we
calculate the timestamp for the enumeration results in
wsman_set_expiretime (wsman-soap-envelope.c), we set it to:

*expire = tv.tv_sec + timeout;

This is basically (epoch seconds + offset).

When we check to see whether the enumeration has expired or not in
wsman_get_expired_enuminfos (wsman-soap.c), we test it against:

mytime = tv.tv_sec * 1000 + tv.tv_usec / 1000;

This is epoch seconds converted to microseconds + epoch microseconds
converted to seconds(?!!). This pretty much always ensures that mytime
will always be greater than expire_time. Thus the enumeration results
expire as soon as they are created.

To fix this, I changed mytime to :

mytime = tv.tv_sec

This seems to work fine. I could not correct the code in
wsman_get_expired_enuminfos, since that code is also used by
subscription_expiration and changing the code there makes
subscriptions act weird.

Tested the code with enumeration intrinsic with expiration time
(PT45S, did not get a chance to test it with datetime format but even
that code seems to convert everything to sec, so I think it should be
OK). Also tested the subscriptions to make sure they were working
correctly.