Found and fixed by Suresh Sundriyal:
Problem:
=====
When you send an enumeration request with an expiretime to Openwsman the
enumerations are not cached and it is immidiately deleted. In order to reproduce
the problem, send an enumeration request to Openwsman with the expiretime
set to 45 seconds. Then try to retrieve the results with the returned enumcontext
and Openwsman returns an invalid enumcontext.

Root cause analysis:
============
The enumerations are tested for expiration in the method wsman_get_expired_enumcontext()
in file wsman-soap.c, where the enumInfo->timeStamp and enumInfo->expiretime are
both tested against the mytime variable which contains the currect epoch time in
milliseconds.

However, the method wsman_set_expiretime() in wsman-soap-envelop.c sets
enumInfo->expireTime in seconds. Thus if enumInfo->expireTime is tested against
mytime, enumInfo->expireTime will always be less. If we change the code such that
mytime is set to current time in epoch seconds (which my previously sent diff did),
then (enumInfo->timeStamp + enumIdletimeout) will always be greater than mytime,
thus breaking normal enumerations without the expiretime.

Fix:
==
As I mentioned before, there were two choices, I could either convert everything to
milliseconds or convert everything to seconds. I decided to go with the latter since:

1) Changing everything to use milliseconds required touching a lot of code, since
subscriptions also use wsman_set_expiretime() and some parts of sfcc-interface
also assume the time to be in seconds.
2) enumInfo->expiretime, enumInfo->timeStamp and mytime are all unsigned long
and on a 32-bit system unsigned long does not have enough bits to store the
current epoch time in milliseconds which as of right now is 1254455904000. (Storing
current epoch time in milliseconds in unsigned long is a bug as well).
3) The cleanup thread runs once every second, so the smallest resolution of time
that you get is a second anyway.

The fix changes mytime to store the current epoch time in seconds. The timestamp
is also set to seconds. The enumIdleTimeout is set to 100 seconds in wsmand-daemon.c
file instead of 100000 milliseconds.