The strdup happens in cim_to_wsman_fault method in sfcc-interface.c. Most of the sfcc wrapper methods in this file call this method and also call each other. There were a couple of places in the file itself where the fault_msg needed to be free'd. While I was at it, I cleaned up the code a bit as well.
The wrapper methods in sfcc-interface.c are called by the callback methods in
cim_data_stubs.c file. Some of the methods in this file are self-contained and
are directly called'back and instantiate their own WsmanStatus variables and
the fault_msg could be free'd from within the methods themselves.
However, methods for Enumerate, Subscribe, Unsubscribe and Renew are called'back from
wrapper methods defined in wsman-soap.c and the WsmanStatus variables are
passed in as pointers and thus needed to be cleaned up in wsman-soap.c.

I hope that makes sense. I have been working on this fix on and off for a while
and it has been sitting in my queue for some time so my recollection of all
the details is a bit fuzzy.

In addition to the fix, it would be nice to document somewhere that in the
future if any additions are made to the CIM plugin files, it should be kept
in mind that the WsmanStatus.fault_msg needs to be explicitly free'd.

Fix by Suresh Sundriyal