OpenWSMAN Client is enabling a host certificate check even if we have not set the verify_host flag.
This is happening due to the eventing portion of the openwsman client
code. In wsman-curl-client-transport.c file at line 335 we can see
that we are checking for NOT NULL of thumbprint and setting
appropriate CURL oprions. Even if we don’t set verify host this option
gets set as thumbprint is not coming as NULL and server is forced to
do a certificate check.

Found and fixed by (A_Venkatachalam@Dell.com)
Reference: bnc#569611