Fix ssl (certificate thumbprint) callback issue
We are seeing an issue where using openwsman client an ssl connection
is not being established even if a valid certificate is available.

This due to a wrong check present in wsman-curl-client-transport.c at
line 336 where we are doing a NULL check for cl->authentication.certificatethumbprint
which is invalid as certificatethumbprint is a array of size 20 of
type char and not a char pointer.

So we must check for length of the string and not a NULL check. Due
to this the callback is set in curl which in-turn sets a callback for
certificate verification in OpenSSL which fails and so does ssl
connection. The callback must not be set as the client has
not set the thumbprint using +wsman_transport_set_certhumbprint()