Fix Windows log functions
The problem (at least with the Visual Studio compiler) is that the
variable arguments defined in the inline functions are not passed into
the debug_full(...) function call. This causes a crash at line 35 of
openwsman-2.2.3/src/lib/u/strings.c when debug_full() calls the helper
function "u_strdup_vprintf()"

This eliminates the redundant inline functions, adds verbose support and
supports multiple debug levels for the WIN32 platform. This compiles and
runs correctly on Visual Studio 2008.

(Found and fixed by Trevor Witten)