Prevent daemon start without valid certificate
Generating a valid certificate requires sufficient entropy for the
random generator. This might not be available in virtualized
environments causing a long delay when reading /dev/random,
effectively preventing openwsmand from starting

This fix removes the automatic creation of the certificate on first
startup and instead instructs the admin to create it manually.

Just ssh-ing to a virtualized server and running the script usually
generates sufficient entropy.