Return current representation after PUT
Using openwsman 2.2.4, a put operation returns the old representation
of the resource (the one sent in the put request message). This occurs
when at least one property is modifiable and a user attempts to set a
non-modifiable property.

DSP0226 specifies that the response shall contain the current
representation of the resource.

In addition, a provider may return an error to indicate a failure to
update one or more of the modifiable properties. However, the "failed"
error code is converted to "action not supported" fault.

- Chris Poblete