Ensure all array elements have same type on invoke
Using openwsman 2.2.4, a segmentation fault occurs when an invalid
reference to a reference array parameter is passed to an invoke
method.  When a reference is not in a valid format, it is marked as a
string.  The proposed fix is for an array parameter, each element is
checked that all are of the same type.

- Chris Poblete