Add indication_baseclass_namespace
From: Till Steinbach [till.steinbach@informatik.haw-hamburg.de]
From: Suresh Sundriyal <ssundriy@vmware.com>

There was a discussion on the list about a year ago regarding subscribing for events from multiple
namespaces via Openwsman. I had looked into this for a while and then promptly forgot about it
and the fix for this issue has kind of bit-rotted on my laptop for some time now.

Recently I received a query regarding this issue from someone else and was reminded of this patch.

With this patch, the semantics for the wsmancli calls are changed such that the user specifies the indication
source namespace instead of the indication profile namespace like he currently has to.

The documentation on the ws-eventing spec is completely useless as to what the namespace in the
subscription message refers to, so I will take it as 'open to interpretation'. The only reference I could find
where this namespace was clarified to be the indication source namespace was in the article provided
by Till (http://www.ibm.com/developerworks/webservices/library/ws-CIM/index.html), so I'm guessing there
are other WS-Management agents out in the wild which use the namespace in the subscription message
to mean indication source namespace. Which would make sense, since the __cimnamespace selector
in the subscription message is not a part of any objectpath/EPR.

Here are my notes from about a year ago (vague as usual and also included in the patch):

'''
While subscribing for indications there are two namespaces involved:

1) The indication profile namespace (where the CIMOM implements the
listener and the handler).
2) The indication source namespace (this is where the indication
provider is registered).

While subscribing for indications with Openwsman, the user can only
specify a single namespace. Currently Openwsman considers this namespace
to be the indication profile namespace for the CIMOM (which is usually
'root/interop' for SFCB) and picks up the indication source namespace
from the configuration file. Problem with this assumption is that the
indication profile namespace never changes whereas the indication source namespace
can change based on where the user has registered his providers. To make
matters worse, a user could have indication providers in multiple
namespaces and then the user would only be able to subscribe to only the
providers that are registered in the namespace that is provided in the
configuration file.

This changes the Openwsman symantics such that the indication profile
namespace is specified in the configuration file (via the
'indication_profile_implementation_ns') and the user can specify the
indication source namespace in his request. This change also obviates
the need for the 'indication_source_namespace' option in the config file
and gets rid of it.

P.S. : This seems like a pretty big shortcoming of the eventing specification.