Place config files is correct directory when CMAKE_INSTALL_PREFIX is used
Installation of config files is failing when PREFIX is specified.

If the build and install is performed by non-root user and PREFIX is
specified, the build will try to install to "/etc".  It fails since
non-root user has no permission.

If CMAKE_INSTALL_PREFIX is /usr, then /etc stays. For any other value
of CMAKE_INSTALL_PREFIX, its prefixed to /etc.