Fix and simplify loading of the Ruby bindings
The Ruby bindings have a binary (.so) part and a Ruby (.rb) part. The
latter provides syntactic sugar for the bindings.
(For Perl and Python, such sugar can be directly added in the .i files)

Before this commit, the Ruby bindings were loaded via
  require 'openwsman/openwsman'
to prevent a conflict of .so and .rb, which are both named openwsman.

This commit renames the .so to _openwsman.so and adds a (toplevel)
openwsman.rb to simplify loading to
  require 'openwsman'

However, Ruby construct the name of the Init_ functions for the
bindings from the required name. And Swig generates the Init_ function
from the module name.

The fix to openwsman.i maps Init__openwsman (the named used by Ruby to
init the binary part of the bindings) to Init_openwsman (the name
generated by Swig).