Handle optional arguments propertly during client endpoint URI creation
The scheme, hostname and path seems to be optional as they are checked
for NULL during assigment to the client data struct. But when creating
the endpoint URI they are used unchecked potentially resulting in a
broken endpoint string or segfault.

Avoid this problem by using the checked values in the client data instead.

Signed-off-by: Klaus Kämpf <kkaempf@suse.de>