Make calls to 'identify' repeatable
Apparently the xml serializer frees data passed from an endpoint stub.
While this is ok for the CIM endpoints, its fatal for the Identify
endpoint since its data is const.
This makes a call to Identify succeed once, all subsequent calls
return empty data.

This commit fixes this by returning a copy of the const data from the
identify stub endpoint and explicitly freeing it in wsman-soap.