Resolve WsXmlDoc.to_xml dilemma
One method provided indented output while keeping the encoding, the
other un-indented output with the ability to change the encoding.
Neither were suitable for WINRM results which are utf-16 encoded.

Luckily, libxml2 provides an method with both desired properties.

This patch makes this method available and fixes WsXmlDoc.to_xml to
use it.