Pass correct key values when creating a Cim_IndicationSubscription
Found and fixed by Santosh_Bidaralli@Dell.com

cim_create_indication_subscription() was adding “OnFatalErrorPolicy,
SubscriptionState, SubscriptionDuration, RepeatNotificationPolicy”
properties as keys to object path, whereas these properties are not the
keys.

The only keys for subscriptions are Filter and Handler and all other
operations related to subscription such as Delete, Update and etc. use
Filter and Handler as the keys.

Actual root cause is that, when a subscription is created the
Subscription hash table inside SFCB was storing the object path as the
key and when unsubscribe request is sent, SFCB tries to look for this
key which does not contain  above mentioned properties in the object path
and hence it fails to find the subscription in its hash table.