curl transport needs option to specify the ssl version used (#29)
libcurl (7.34 or newer) seems to try to negotiate tlsv1.1/tlsv1.2 which causes problems at least with Intel AMT 8. You will get a "sslv3 alert bad record mac" and thus can't connect via SSL/TLS which is bad in case the AMT device is configured for TLS only access.
You need to explicitely specify sslv3 or tlsv1.0 to be able to connect to Intel AMT.
The patch below adds an environment variable, that, if set, enforces the use of a specific ssl or tls version.
```

Signed-off-by: Klaus Kämpf <kkaempf@suse.de>