Allow EndpointReference in ClientOptions properties
In order to support passing endpoint references as parameters of
invoke actions, the wsmc_add_property call needs to be extended.

In the current implementation, only string values can be added as
option properties.

This commit add a new wsmc_add_property_epr() call to add a non-string
'epr_t *' property.
Internally, such properties are encoded with an '<epr>' prefix in the
key, serving as type information for the data.
Also the hash cleanup callback is modified to properly free either
string or epr_t data.