Added new config option 'cim_client_cql' to specify CQL query identifier
Dafaults to 'CQL' to be backwards-compatible with earlier releases of
Openwsman.
Should be set to 'DMTF:CQL' for better standards-conformance, e.g.
with tog-pegasus 2.13 and later.

DPS0202 (CIM Query Language Specification), released in 2007,
specifies the identifier for CQL (CIM Query Language) to 'DMTF:CQL'.

However, Openwsman used 'CQL' which was working fine with sblim-sfcb
and tog-pegasus < 2.13.

Apparently, top-pegasus 2.13 changed this to be more strict, while
sblim-sfcb > 1.4 accepts 'CQL' and 'DMTF:CQL'.

This fixes issue #60